+++ setup-after-install.sh
@@ -4,9 +4,9 @@ echo "Prepare husky..."
 pnpm husky install
 
 echo "Setup environment variables..."
-cp apps/things/assets-catalog/rest-api/.env.example apps/things/assets-catalog/rest-api/.env
-cp apps/kernel/taxonomic-units/rest-api/.env.example apps/kernel/taxonomic-units/rest-api/.env
-cp apps/people/organizations-management/rest-api/.env.example apps/people/organizations-management/rest-api/.env
+cp teams/things/assets-catalog/rest-api/.env.example teams/things/assets-catalog/rest-api/.env
+cp teams/kernel/taxonomic-units/rest-api/.env.example teams/kernel/taxonomic-units/rest-api/.env
+cp teams/people/organizations-management/rest-api/.env.example teams/people/organizations-management/rest-api/.env
 
 echo "Generating database clients..."
 pnpm prisma generate --schema libs/researchers/peers/adapters/src/database/infra/prisma/postgresql.schema.prisma
@@ -15,7 +15,7 @@ echo "Building app..."
 pnpm nx run-many --target=build --all --skip-nx-cache
 
 echo "Copying .env.example..."
-cp apps/things/assets-catalog/rest-api/.env.example dist/apps/things/assets-catalog/rest-api/.env
+cp teams/things/assets-catalog/rest-api/.env.example dist/teams/things/assets-catalog/rest-api/.env
 cp .env.example .env
 
 echo "Creating log file..."